Compile-time gcc warning due to unsigned conversion

+++ test/core/testCoordinate.cpp
@@ -9,14 +9,14 @@ TEST(TestCoordinate, PSCoordinate)
 {
     /* constructor test */
     PS::Coordinate a{15, 25};
-    EXPECT_EQ(15, a.x);
-    EXPECT_EQ(25, a.y);
+    EXPECT_EQ(static_cast<unsigned>(15), a.x);
+    EXPECT_EQ(static_cast<unsigned>(25), a.y);
 
     /* setter test */
     a.x = 35;
     a.y = 52;
-    EXPECT_EQ(35, a.x);
-    EXPECT_EQ(52, a.y);
+    EXPECT_EQ(static_cast<unsigned>(35), a.x);
+    EXPECT_EQ(static_cast<unsigned>(52), a.y);
 }
 
 TEST(TestCoordinate, PSRGBAColor)
@@ -39,7 +39,7 @@ TEST(TestCoordinate, PSRGBAColor)
     EXPECT_EQ(uint32_t(0x56), a.blue());
 
     /* Test aggregate size */
-    EXPECT_EQ(4, sizeof(a));
+    EXPECT_EQ(unsigned(4), sizeof(a));
 }
 
 TEST(TestCoordinate, PSPixel)
@@ -47,8 +47,8 @@ TEST(TestCoordinate, PSPixel)
     /* Test aggregate initializer */
     PS::Pixel a{15, 25, 0xDEADCAFE};
     EXPECT_EQ(0xDEADCAFE, a.color);
-    EXPECT_EQ(15, a.x);
-    EXPECT_EQ(25, a.y);
+    EXPECT_EQ(static_cast<unsigned>(15), a.x);
+    EXPECT_EQ(static_cast<unsigned>(25), a.y);
 
     /* Test getters */
     EXPECT_EQ(uint32_t(0xAD), a.red());
@@ -64,10 +64,10 @@ TEST(TestCoordinate, PSPixel)
     EXPECT_EQ(uint32_t(0x12), a.red());
     EXPECT_EQ(uint32_t(0x34), a.green());
     EXPECT_EQ(uint32_t(0x56), a.blue());
-    EXPECT_EQ(10, a.x);
-    EXPECT_EQ(20, a.y);
+    EXPECT_EQ(static_cast<unsigned>(10), a.x);
+    EXPECT_EQ(static_cast<unsigned>(20), a.y);
 
     /* Test aggregate size */
-    EXPECT_EQ(12, sizeof(a));
+    EXPECT_EQ(unsigned(12), sizeof(a));
 }
